move ids to separate file
