+++ packages/frontend/src/components/project/navigation/index.ts
@@ -1,0 +1,15 @@
+export const DESKTOP_PROJECT_NAVIGATION_IDS = {
+  container: 'desktop-project-navigation',
+  listHeader: 'desktop-project-navigation-list-header',
+  list: 'desktop-project-navigation-list',
+  summaryItem: 'desktop-project-navigation-summary-item',
+  index: 'desktop-project-navigation-item-index',
+}
+
+export const MOBILE_PROJECT_NAVIGATION_IDS = {
+  container: 'mobile-project-navigation',
+  list: 'mobile-project-navigation-list',
+  summaryItem: 'mobile-project-navigation-summary',
+  arrowLeft: 'mobile-project-navigation-arrow-left',
+  arrowRight: 'mobile-project-navigation-arrow-right',
+}
